refactor: Put json extension when requiring locales
Unfortunately, this does not work properly as require does not create
a context where only JSON files are imported. This causes problems when
a .DS_Store file is in the locales directory. Only solution now is to
delete the .DS_Store file from the locales directory.

At least now, it is more explicit that we want only the JSON files in
the locales directory to be included in the build. To me, the bug lies
in webpack now (it should automatically filter the context when the
extension is there).