fix: Remove vertical scrolling on mobile menu
Overwriting the ::after element in Layout because this page
is a special case and doesn't have anything to show on the bottom
of the page. So we just remove that element which was creating
unwanted scrolling. At this point I didn't think updating cozy-ui
was necessary because it is a special app scenario