refactor: Make color scheme select react to update from other tab/device
When on cozy-settings with the select set on 'light' , we want that if we change to 'dark' on another tab/device, the value is changed on the first tab/device.

The more consise way is to remove the state : displayed value comes directly from the instance setting coming from a query in a parent component (which receives update from realtime and save).