fix: Enforce stronger encryption for new password hashing
Previously, we attempted to hash the new password using the same encryption parameters as the initial password. In cases where the initial password was hashed with low kdf iterations, this led to server rejections. Thus commit fore a higher kdf iterations for new passwords to ensure compatibility with server requirements