feat: Update features displayed in plan and offer page
For some features, it is just an icon or wording changes.

For other features, we changed the way we display them :  we do
not display anymore disabled feature but display instead a
default value like 'Passwords manager' for passwords, and 'Passwords
manager with sharing' when the feature is enabled.

I avoided the generic SubscriptionFlagItem because there is always
a small difference in the way we get the icon or wording. So the generic
way which is already complicated would have been a lot more complicated.