feat: New email only method to delete account
Because we are not ready to delete Cozy and Twake account
automatically, we add a new email only method to delete account. When
the flag `settings.delete.byEmailOnly` is enabled, it does not delete
the account but instead send an email to the support address to ask
the account deletion.